+++ filesStructure.md
@@ -1,0 +1,25 @@
+src/
+├── assets/
+│   ├── *.svg
+│   └── *.wav
+├── components/
+│   ├── AddButton.jsx
+│   ├── CircularIndeterminate.jsx
+│   ├── Modal.jsx
+│   ├── Pagination.jsx
+│   ├── RemoveButton.jsx
+│   └── Timer.jsx
+├── pages/
+│   ├── Login/
+│   │   └── Login.jsx
+│   └── WorkoutPage/
+│       ├── EditExercisesForm.jsx
+│       ├── Exercise.jsx
+│       ├── Workout.jsx
+│       └── WorkoutForm.jsx
+├── App.jsx
+├── AuthContext.jsx
+├── firebaseConfig.js
+├── index.css
+└── main.jsx
+
