+++ src/index.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Itim&display=swap');
 
 body {
-    background-color: #7F00CD !important;
+    background-color: #1d1680 !important;
 }
 
 .app {
