+++ src/testFunctions.js
@@ -1,0 +1,40 @@
+describe('Login Component', () => {
+    test('CT001 - Login bem-sucedido', async () => {
+        // Arrange
+        render(<Login />);
+        const emailInput = screen.getByTestId('email-input');
+        const passwordInput = screen.getByTestId('password-input');
+
+        // Act
+        fireEvent.change(emailInput, { target: { value: 'test@test.com' } });
+        fireEvent.change(passwordInput, { target: { value: 'password123' } });
+        fireEvent.click(screen.getByText('Entrar'));
+
+        // Assert
+        expect(await screen.findByText('Dashboard')).toBeInTheDocument();
+    });
+
+    test('CT002 - Login com credenciais inválidas', async () => {
+        // Similar ao anterior, testando erro
+    });
+});
+
+describe('WorkoutForm Component', () => {
+    test('CT003 - Criação de novo treino', async () => {
+        // Teste de criação de treino
+    });
+
+    test('CT004 - Edição de treino existente', async () => {
+        // Teste de edição de treino
+    });
+});
+
+describe('Timer Component', () => {
+    test('CT005 - Iniciar timer', () => {
+        // Teste de início do timer
+    });
+
+    test('CT006 - Parar timer', () => {
+        // Teste de parada do timer
+    });
+});
